requirements.txt
torch
numpy
gymnasium
matplotlib
tqdm
pyyaml
tetris-gymnasium
ale-py
